+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-05-14 12:05:48
-// Commit: feat(forms): add bi-directional event bus handler for forms
+// Updated at: 2026-05-14 17:39:25
+// Commit: chore(forms): update dependencies and cleanup unused imports in forms
 
 class formsComponentState {
-  final String commitRef = '1134';
-  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
-  final DateTime timestamp = DateTime.parse('2026-05-14T12:05:48+05:00');
+  final String commitRef = '1154';
+  final String lastMessage = 'chore(forms): update dependencies and cleanup unused imports in forms';
+  final DateTime timestamp = DateTime.parse('2026-05-14T17:39:25+05:00');
 }